FileTypeToolkit merges modomics with Andrew data
The code is horrendous and uncommented, but in FileTypeToolkit.py
is how I added the Modomics true positives to the Sacc Cere data.

The only issues are that I cannot add modomics data to modifications
we havent run with the pipeline, and it takes a pretty long time (180 s)
to generate the isItSmaller.csv file.

I have a complexity x^2 for my list searches so that is why the time is
pretty awful.

Whenever the database expands I think we will need to rerun this in
order to add in the extra modomics data (will definitely need to find
a faster way of doing this)